Add: Grunt workflow
This change adds a gruntfile which automates several steps of the
development workflow. It also removes the custom module wrapper
that provided browser compat in favour of browserify via
grunt-browserify.

While nothing related to the Defer API was changed the version is
still getting a bump to reflect the fairly significant change in
methodology for cross-platform support.

Signed-off-by: Kevin Conway <kevinjacobconway@gmail.com>